+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.3] - 2019-10-09
+## [0.2.3] - 2019-10-12
 
 ### Added
 
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add property `title` to Line-Chart
 - Add property `title` to Scatter-Chart
 - Add property `title` to Pie-Chart
+- Add properties `xlabel` and `ylabel` to Line-Chart
 
 ## [0.2.2] - 2019-10-06